We don't want to convert dates during upgrade anymore.
Instead, we're going to parse/save dates into the app's timezone.
