+++ src/Providers/ExtensionServiceProvider.php
@@ -246,7 +246,7 @@ class ExtensionServiceProvider extends ServiceProvider
         Updates\AddSitePermissions::class,
         Updates\UseClassBasedStatamicUniqueRules::class,
         Updates\MigrateSitesConfigToYaml::class,
-        Updates\ConvertDatesToUtc::class,
+        Updates\AddTimezoneOptionsToSystemConfig::class,
         Updates\RemoveParentField::class,
     ];
